ゲーム内の時間構造を別ファイルで宣言する
